Try to parse with old CLI initially, fall back to the new one Move conversion logic to TryFrom Add help-deprecated command Document usages of old and new cli
